% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble_package.R
\name{pred_new}
\alias{pred_new}
\title{Caret ensemble list prediction}
\usage{
pred_new(i, model_list, data)
}
\arguments{
\item{i}{model index}

\item{model_list}{caret list containing list of models}

\item{data}{data on which prediction should be performed}
}
\value{
Vector of predictions for p(y=1)
}
\description{
This function takes the model list and runs the prediction for one model
}
\examples{
pred_new(i = 1, model_list = caretlist, data = train_data)

}
