% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble_package.R
\name{threshold2}
\alias{threshold2}
\title{Threshold calculation}
\usage{
threshold2(predict, response, type = "sens+spec")
}
\arguments{
\item{predict}{Vector of predicted p(y=1)}

\item{response}{Vector of actual class labels}

\item{type}{(Optional) Threshold type. Currently "ppv" and "sens+spec"
(default) are supported}
}
\value{
Threshold value
}
\description{
This function calculates threshold based on given condition
}
\examples{
threshold2(response = c(0,1), predict = c(0.1, 0.9))

}
